requirements.txt
beautifulsoup4>=4.12.0
PyYAML>=6.0
requests>=2.31.0
anthropic>=0.39.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
jieba>=0.42.0
